Finished InstructionDecoder tests and added test for shared mem program.
