+++ src/Solnet.Examples/TransactionBuilderExample.cs
@@ -9,6 +9,7 @@ using Solnet.Wallet;
 using Solnet.Wallet.Utilities;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Threading;
