+++ test/Solnet.Programs.Test/Resources/TestDecodeFromTransactionUnknownInstruction.json
@@ -1,0 +1,178 @@
+{
+  "blockTime": 1628189179,
+  "meta": {
+    "err": null,
+    "fee": 10000,
+    "innerInstructions": [
+      {
+        "index": 2,
+        "instructions": [
+          {
+            "accounts": [
+              3,
+              1,
+              5
+            ],
+            "data": "3DaiBu8ogoUK",
+            "programIdIndex": 11
+          }
+        ]
+      }
+    ],
+    "logMessages": [
+      "Program 11111111111111111111111111111111 invoke [1]",
+      "Program 11111111111111111111111111111111 success",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA invoke [1]",
+      "Program log: Instruction: InitializeAccount",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA consumed 4428 of 200000 compute units",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA success",
+      "Program auctxRXPeJoc4817jDhf4HbjnhEcr1cCXenosMhK5R8 invoke [1]",
+      "Program log: + Processing Cancelbid",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA invoke [2]",
+      "Program log: Instruction: Transfer",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA consumed 3513 of 176043 compute units",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA success",
+      "Program auctxRXPeJoc4817jDhf4HbjnhEcr1cCXenosMhK5R8 consumed 32013 of 200000 compute units",
+      "Program auctxRXPeJoc4817jDhf4HbjnhEcr1cCXenosMhK5R8 success",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA invoke [1]",
+      "Program log: Instruction: CloseAccount",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA consumed 2322 of 200000 compute units",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA success"
+    ],
+    "postBalances": [
+      29560522283,
+      0,
+      1566000,
+      2039280,
+      1454640,
+      4217760,
+      2415120,
+      113939664699,
+      1,
+      1,
+      1,
+      1130582400,
+      1141440
+    ],
+    "postTokenBalances": [
+      {
+        "accountIndex": 3,
+        "mint": "So11111111111111111111111111111111111111112",
+        "uiTokenAmount": {
+          "amount": "0",
+          "decimals": 9,
+          "uiAmount": null,
+          "uiAmountString": "0"
+        }
+      }
+    ],
+    "preBalances": [
+      23560532283,
+      0,
+      1566000,
+      6002039280,
+      1454640,
+      4217760,
+      2415120,
+      113939664699,
+      1,
+      1,
+      1,
+      1130582400,
+      1141440
+    ],
+    "preTokenBalances": [
+      {
+        "accountIndex": 3,
+        "mint": "So11111111111111111111111111111111111111112",
+        "uiTokenAmount": {
+          "amount": "6000000000",
+          "decimals": 9,
+          "uiAmount": 6.0,
+          "uiAmountString": "6"
+        }
+      }
+    ],
+    "rewards": [],
+    "status": {
+      "Ok": null
+    }
+  },
+  "slot": 90284855,
+  "transaction": {
+    "message": {
+      "accountKeys": [
+        "hoakwpFB8UoLnPpLC56gsjpY7XbVwaCuRQRMQzN5TVh",
+        "3ejUMZigjDD2kKMT4HZkEa6mgZWX1gAmjqcioTJQosBg",
+        "3c58FLuuqbW6L894Ld6J7mnFtxj9oSkfejEhjm6HUpZ8",
+        "C9KnqWGq6wEApwAuGV4Be8oMGuxCpmgi5p4FcVN8QdGa",
+        "2sBWKnucDTSC3HFJMtrB8otfyrvfunzn7mr3t2j2j8mx",
+        "9VPY2BVoMCUpCZSRecfKTs6Mxx2f5VWbA1oGKtSxVnvN",
+        "AycprntWgDGw2YKi1q9o4uDX5PVp9Wg8d9rRAN9VHJ3E",
+        "So11111111111111111111111111111111111111112",
+        "SysvarRent111111111111111111111111111111111",
+        "SysvarC1ock11111111111111111111111111111111",
+        "11111111111111111111111111111111",
+        "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA",
+        "auctxRXPeJoc4817jDhf4HbjnhEcr1cCXenosMhK5R8"
+      ],
+      "header": {
+        "numReadonlySignedAccounts": 0,
+        "numReadonlyUnsignedAccounts": 5,
+        "numRequiredSignatures": 2
+      },
+      "instructions": [
+        {
+          "accounts": [
+            0,
+            1
+          ],
+          "data": "11119os1e9qSs2u7TsThXqkBSRVFxhmYaFKFZ1waB2X7armDmvK3p5GmLdUxYdg3h7QSrL",
+          "programIdIndex": 10
+        },
+        {
+          "accounts": [
+            1,
+            7,
+            0,
+            8
+          ],
+          "data": "2",
+          "programIdIndex": 11
+        },
+        {
+          "accounts": [
+            0,
+            1,
+            2,
+            3,
+            4,
+            5,
+            6,
+            7,
+            9,
+            8,
+            10,
+            11
+          ],
+          "data": "1Cn4ibidHMiHSswjcjiqxWsbVjuTZQj3VxmPaLxkrxL5b",
+          "programIdIndex": 12
+        },
+        {
+          "accounts": [
+            1,
+            0,
+            0
+          ],
+          "data": "A",
+          "programIdIndex": 11
+        }
+      ],
+      "recentBlockhash": "7abQTYyiPRgWQ4bapq7ct4JsGVBNFwjFYRFV5AN4ap2c"
+    },
+    "signatures": [
+      "5AB4XLG2CNX3ShWHoLseaZva26ygsPpKxUNHKLq9ijQsQPVJovg2yjQ5n25y2BjXbZE7Qid1K3sxB3MNaB2BkjAX",
+      "5LFCkF1M8HiawFkmAWECQXDJva6X5PGLGgaRUfWdh2tVSezn4gdHRztxvnHBLGFra85zcQz1ienpmiw29J1Fc9rw"
+    ]
+  }
+}
